requirements.txt
flask-restx
Flask-Migrate
pyjwt
flask_testing
Flask-Script
flask-bcrypt
flask-cors